Create Countdown
Code that simulates a game of Countdown. Letter and numbers are in their collections in the same ratio as the tv series.
The numbers game takes the desired number of big numbers as an argument (of if left blank takes user input to choose how many big numbers). It generates a list of 6 numbers, starting with the big numbers, then generates a target number and returns both.

TODO: Implement the letters game
TODO: Add a 30 second timer