Create tweetJumbler.py
Code pulls a large number of a user's tweets, places each word into a dictionary with every word that follows it in the tweets in a list as the value. The script then uses a random follow-up word from a random initial word to construct a full tweet that rarely if ever makes sense.